Added Login.java, Payment.java, Reservation.java, added a .db file to hold all the data tables.
