+++ src/Payment.java
@@ -1,0 +1,5 @@
+public class Payment {
+    private String cardType;
+    private int cardNum;
+    private String bank;
+}
